requirements.txt
docx==0.2.4
pandas==1.2.4
Pillow==9.0.1
python_docx==0.8.11
streamlit==1.4.0
